refactor: split E2E tests by test classification
Split toggle-rule-active.spec.ts into separate files per test classification:
- helpers.ts: Common helper functions (saveRule, getToggleState, clickToggle)
- normal-cases.spec.ts: Normal operation flow + DB persistence
- data-independence.spec.ts: Multiple data independence
- concurrency.spec.ts: Concurrency prevention

Update guideline and test strategy to reflect file-per-classification pattern.
