+++ docs/design/pages/rule-list/features/toggle-rule-active/e2e-test-strategy.md
@@ -32,28 +32,19 @@
 
 ## テスト分類
 
-### 1. 正常操作フロー - 状態切り替え
+### 1. 正常操作フロー - 状態切り替え + DB永続化
 
-トグルスイッチの基本操作が正しく動作することを検証する。
+トグルスイッチの基本操作が正しく動作し、状態がDBに永続化されることを検証する。
 
 | 分類 | テストケース | 根拠 |
 |------|-------------|------|
 | true → false | 有効なルールを無効に切り替え | 00-overview.md「有効/無効トグル」の基本パターン |
 | false → true | 無効なルールを有効に切り替え | 00-overview.md「有効/無効トグル」の基本パターン |
-
-**対応テスト**: `toggle-rule-active.spec.ts`
-
-### 2. 状態永続化 - DB保存
-
-トグル状態がDBに正しく保存され、リロード後も維持されることを検証する。
-
-| 分類 | テストケース | 根拠 |
-|------|-------------|------|
 | リロード後維持 | ページリロード後もトグル状態が維持される | 00-overview.md「変更をDBに保存する」 |
 
-**対応テスト**: `toggle-rule-active.spec.ts`
+**対応テスト**: `normal-cases.spec.ts`
 
-### 3. 複数データ操作 - 独立性
+### 2. 複数データ操作 - 独立性
 
 複数ルールがある場合に、操作対象以外のルールに影響がないことを検証する。
 
@@ -61,9 +52,9 @@
 |------|-------------|------|
 | 他ルール不変 | 1つのルールをトグルしても他ルールは変化しない | データ独立性の保証 |
 
-**対応テスト**: `toggle-rule-active.spec.ts`
+**対応テスト**: `data-independence.spec.ts`
 
-### 4. 競合防止
+### 3. 競合防止
 
 同一ルールへの連続操作が適切に制御されることを検証する。
 
@@ -71,7 +62,7 @@
 |------|-------------|------|
 | 操作完了 | トグル操作が正常に完了する | 00-overview.md「競合状態の防止」 |
 
-**対応テスト**: `toggle-rule-active.spec.ts`
+**対応テスト**: `concurrency.spec.ts`
 
 ## ユーザーストーリートレーサビリティ
 
@@ -79,14 +70,14 @@
 
 | 機能要件 | テストケース | テストファイル |
 |---------|-------------|---------------|
-| トグル操作でルールの有効/無効を切り替える | true→false、false→true切り替え | toggle-rule-active.spec.ts |
-| 変更をDBに保存する | リロード後もトグル状態が維持される | toggle-rule-active.spec.ts |
+| トグル操作でルールの有効/無効を切り替える | true→false、false→true切り替え | normal-cases.spec.ts |
+| 変更をDBに保存する | リロード後もトグル状態が維持される | normal-cases.spec.ts |
 
 ### 競合状態の防止（00-overview.md参照）
 
 | 要件 | テストケース | テストファイル |
 |-----|-------------|---------------|
-| トグル処理中はトグルスイッチを操作不可にする | トグル操作が正常に完了する | toggle-rule-active.spec.ts |
+| トグル処理中はトグルスイッチを操作不可にする | トグル操作が正常に完了する | concurrency.spec.ts |
 
 ### E2Eテストスコープ外
 
@@ -111,7 +102,10 @@
 
 ```
 tests/e2e/pages/rule-list/features/toggle-rule-active/
-└── toggle-rule-active.spec.ts    # トグル機能E2Eテスト
+├── helpers.ts                  # 共通ヘルパー関数
+├── normal-cases.spec.ts        # 正常操作フロー + DB永続化
+├── data-independence.spec.ts   # 複数データ独立性
+└── concurrency.spec.ts         # 競合防止
 ```
 
 ## テスト環境
