+++ docs/docs-rules/design/07-e2e-test-strategy.md
@@ -71,7 +71,7 @@ E2Eテスト特有の検証観点:
 
 00-overview.mdの機能要件を踏まえて分類する。
 
-各テスト分類には対応するテストファイルを明記する:
+**テスト分類ごとにファイルを分割する**。各テスト分類に対応するテストファイルを作成し、テストの責務を明確にする。
 
 ```markdown
 ### 1. 正常操作フロー - 状態切り替え
@@ -83,7 +83,29 @@ E2Eテスト特有の検証観点:
 | true → false | 有効なルールを無効に切り替え | 基本パターン |
 | false → true | 無効なルールを有効に切り替え | 基本パターン |
 
-**対応テスト**: `toggle-rule-active.spec.ts`
+**対応テスト**: `normal-cases.spec.ts`
+
+### 2. 状態永続化 - DB保存
+
+[この観点でテストする理由]
+
+| 分類 | テストケース | 根拠 |
+|------|-------------|------|
+| リロード後維持 | ページリロード後もトグル状態が維持される | DB永続化 |
+
+**対応テスト**: `persistence.spec.ts`
+```
+
+### テストファイル構成
+
+テスト分類ごとにファイルを分割して配置する:
+
+```
+tests/e2e/pages/{page-name}/features/{feature-name}/
+├── normal-cases.spec.ts        # 正常操作フロー
+├── persistence.spec.ts         # 状態永続化
+├── data-independence.spec.ts   # 複数データ独立
+└── error-cases.spec.ts         # エラー系（該当する場合）
 ```
 
 ## 記述ルール
