+++ host-frontend-root/frontend-src-root/tests/e2e/pages/rule-list/features/toggle-rule-active/normal-cases.spec.ts
@@ -1,0 +1,102 @@
+import { expect, test } from 'tests/e2e/fixtures';
+import { clickToggle, getToggleState, saveRule } from 'tests/e2e/pages/rule-list/features/toggle-rule-active/helpers';
+
+/**
+ * ルールトグル機能 - 正常操作フローのE2Eテスト
+ *
+ * トグルスイッチの基本操作が正しく動作し、状態がDBに永続化されることを検証する。
+ *
+ * @see docs/design/pages/rule-list/features/toggle-rule-active/e2e-test-strategy.md
+ */
+test.describe('ルールトグル機能 - 正常操作フロー', () => {
+  test('有効なルールを無効に切り替えられる（isActive: true → false）', async ({ page, popupPage, rulesPage }) => {
+    // コンソールエラーメッセージを記録するための配列
+    const consoleMessages: string[] = [];
+
+    rulesPage.on('console', msg => {
+      if (msg.type() === 'error') {
+        consoleMessages.push(msg.text());
+      }
+    });
+
+    // 1. Arrange: ルールを保存（デフォルトでisActive=true）
+    await saveRule(popupPage, page, {
+      oldString: 'テスト文字列A',
+      newString: '置換後文字列A',
+    });
+
+    // 2. Arrange: ルール一覧ページをリロード
+    await rulesPage.reload();
+    await expect(rulesPage.locator('[data-testid="rules-table"]')).toBeVisible({ timeout: 60000 });
+
+    // 3. Assert: 初期状態でトグルがON（有効）であることを確認
+    const initialState = await getToggleState(rulesPage, 0);
+    expect(initialState).toBe(true);
+
+    // 4. Act: トグルをクリックして無効に切り替え
+    await clickToggle(rulesPage, 0);
+
+    // 5. Assert: トグルがOFF（無効）になったことを確認
+    await expect(async () => {
+      const newState = await getToggleState(rulesPage, 0);
+      expect(newState).toBe(false);
+    }).toPass({ timeout: 10000 });
+
+    // 6. Assert: ページをリロードしてもトグル状態が維持されていることを確認（DB永続化）
+    await rulesPage.reload();
+    await expect(rulesPage.locator('[data-testid="rules-table"]')).toBeVisible({ timeout: 60000 });
+
+    const persistedState = await getToggleState(rulesPage, 0);
+    expect(persistedState).toBe(false);
+
+    // 7. Assert: コンソールエラーが発生していないことを確認
+    expect(consoleMessages).toHaveLength(0);
+  });
+
+  test('無効なルールを有効に切り替えられる（isActive: false → true）', async ({ page, popupPage, rulesPage }) => {
+    // コンソールエラーメッセージを記録するための配列
+    const consoleMessages: string[] = [];
+
+    rulesPage.on('console', msg => {
+      if (msg.type() === 'error') {
+        consoleMessages.push(msg.text());
+      }
+    });
+
+    // 1. Arrange: ルールを保存
+    await saveRule(popupPage, page, {
+      oldString: 'テスト文字列B',
+      newString: '置換後文字列B',
+    });
+
+    // 2. Arrange: ルール一覧ページをリロード
+    await rulesPage.reload();
+    await expect(rulesPage.locator('[data-testid="rules-table"]')).toBeVisible({ timeout: 60000 });
+
+    // 3. Arrange: トグルをOFFにする（無効化）
+    await clickToggle(rulesPage, 0);
+    await expect(async () => {
+      const state = await getToggleState(rulesPage, 0);
+      expect(state).toBe(false);
+    }).toPass({ timeout: 10000 });
+
+    // 4. Act: トグルをクリックして有効に切り替え
+    await clickToggle(rulesPage, 0);
+
+    // 5. Assert: トグルがON（有効）になったことを確認
+    await expect(async () => {
+      const newState = await getToggleState(rulesPage, 0);
+      expect(newState).toBe(true);
+    }).toPass({ timeout: 10000 });
+
+    // 6. Assert: ページをリロードしてもトグル状態が維持されていることを確認（DB永続化）
+    await rulesPage.reload();
+    await expect(rulesPage.locator('[data-testid="rules-table"]')).toBeVisible({ timeout: 60000 });
+
+    const persistedState = await getToggleState(rulesPage, 0);
+    expect(persistedState).toBe(true);
+
+    // 7. Assert: コンソールエラーが発生していないことを確認
+    expect(consoleMessages).toHaveLength(0);
+  });
+});
